+++ instapy/story_util.py
@@ -1,8 +1,12 @@
+# import built-in & third-party modules
 import time
 import math
+import requests
+
 from random import randint
-from selenium.common.exceptions import NoSuchElementException
-from selenium.common.exceptions import WebDriverException
+
+
+# import InstaPy modules
 from .util import click_element
 from .util import web_address_navigator
 from .util import update_activity
@@ -10,7 +14,9 @@ from .util import get_action_delay
 from .settings import Settings
 from .xpath import read_xpath
 
-import requests
+# import exceptions
+from selenium.common.exceptions import NoSuchElementException
+from selenium.common.exceptions import WebDriverException
 
 
 def get_story_data(browser, elem, action_type, logger, simulate=False):